requirements.txt
openai>=1.0.0
flask>=2.0.0
gunicorn>=21.0.0
python-dotenv>=0.21.0
requests>=2.28.0
